dependencies: refactor python3 into its own file
